MJ02 - Sandbox cleanup, Pinger test version
DustSandbox created to collect all "sandbox" component initialization;
modules and call location updated.

Created a sandbox/http package with an initial "Pinger" implementation.
This should grow to a Master (running a Jetty server, responds to a call
with "Single" JSON containing all changed entities) and a Slave (regular
ping to the server, load and update the knowledge).